+++ packages/components/textarea/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @jamsr-ui/input
 
+## 0.0.82
+
+### Patch Changes
+
+- fixes
+- Updated dependencies
+  - @jamsr-ui/input@0.0.82
+
+## 0.0.81
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/input@0.0.81
+
 ## 0.0.80
 
 ### Patch Changes